// src/types.ts
type Ref<T={}> = Signal<T>
// ...
export interface IResponseError {
    errorCode?: string;
    fieldName?: string;
    message?: string;
}

export interface IResponseStatus extends IResponseError {
    errors?: ResponseError[];
}

export type ApiState = {
    unRefs: (o:any) => any
    setRef: ($ref:Ref<any>, $item:any) => void
    setError: ({ message, errorCode, fieldName, errors }: IResponseStatus) => ResponseStatus
    addFieldError: ({ fieldName, message, errorCode }: IResponseError) => void
    loading: Ref<boolean>
    error: Ref<any>
    api: <TResponse>(request: IReturn<TResponse> | ApiRequest, args?: any, method?: string) => Promise<ApiResult<TResponse>>
    apiVoid: (request: IReturnVoid | ApiRequest, args?: any, method?: string) => Promise<ApiResult<EmptyResponse>>
    apiForm: <TResponse>(request: ApiRequest | IReturn<TResponse>, body: FormData, args?: any, method?: string) => Promise<ApiResult<TResponse>>
    apiFormVoid: (request: IReturnVoid | ApiRequest, body: FormData, args?: any, method?: string) => Promise<ApiResult<EmptyResponse>>
    swr: <TResponse>(request:IReturn<TResponse> | ApiRequest, fn:(r:ApiResult<TResponse>) => void, args?: any, method?: string) => Promise<ApiResult<TResponse>>
    swrEffect: <TResponse>(requestFn: () => IReturn<TResponse> | ApiRequest, options?:{ args?:any, method?:string, delayMs?:number }) => Ref<ApiResult<TResponse>>
}
// ...
export interface ApiResponse {
    response?: any;
    error?: ResponseStatus;
    get completed(): boolean;
    get failed(): boolean;
    get succeeded(): boolean;
    get errorMessage(): string;
    get errorCode(): string;
    get errors(): ResponseError[];
    get errorSummary(): string;
}
export interface ApiResult<TResponse> extends ApiResponse {
    response?: TResponse;
    error?: ResponseStatus;
    get completed(): boolean;
    get failed(): boolean;
    get succeeded(): boolean;
    get errorMessage(): string;
    get errorCode(): string;
    get errors(): ResponseError[];
    get errorSummary(): string;
    fieldError(fieldName: string): ResponseError;
    fieldErrorMessage(fieldName: string): string;
    hasFieldError(fieldName: string): boolean;
    showSummary(exceptFields?: string[]): boolean;
    summaryMessage(exceptFields?: string[]): string;
    addFieldError(fieldName: string, message: string, errorCode?: string): void;
}

export interface ApiRequest {
    [k:string]: any;
    getTypeName(): string;
    getMethod(): string;
    createResponse(): any;
}
export interface IReturnVoid {
    createResponse(): any;
}
export interface IReturn<T> {
    createResponse(): T;
}
export interface ResponseStatus {
    errorCode?: string;
    message?: string;
    stackTrace?: string;
    errors?: ResponseError[];
    meta?: {
        [index: string]: string;
    };
}
export interface ResponseError {
    errorCode?: string;
    fieldName?: string;
    message?: string;
    meta?: {
        [index: string]: string;
    };
}
export interface ErrorResponse {
    responseStatus?: ResponseStatus;
}
export interface EmptyResponse {
    responseStatus?: ResponseStatus;
}